pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "tick"
version = "0.0.0"
dependencies = [
	"pyjwt",
	"pytest",
	"python-dotenv",
	"pytz",
	"requests",
	"sqlalchemy",
	"tqdm",
]
